# starmap-contracts/openapi.yaml
openapi: "3.0.3"
info:
  title: "星图 StarMap API"
  description: "人才能力星云导航系统 - 后端 API 契约"
  version: "1.0.0"
  contact:
    name: "StarMap Team"
  x-summary: "人才技能图谱知识系统全量 API"
servers:
  - url: "http://localhost:8000/api/v1"
    description: "本地开发环境"
  - url: "https://api.starmap.dev/api/v1"
    description: "评审环境"
security:
  - BearerAuth: []
paths:
...
  /admin/prompts/{name}:
    get:
      operationId: getPromptInfo
      summary: "提示词模板详情"
      description: "获取指定提示词模板的元数据"
      tags: [管理后台]
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "模板详情"
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "404":
          description: "模板不存在"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/prompts/{name}/template:
    get:
      operationId: getPromptTemplateContent
      summary: "提示词模板内容"
      description: "获取提示词模板的原始内容"
      tags: [管理后台]
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "模板内容"
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                  template:
                    type: string
  /admin/prompts/{name}/versions:
    post:
      operationId: createPromptVersion
      summary: "注册提示词版本"
      description: "注册新的提示词版本"
      tags: [管理后台]
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [template]
              properties:
                template:
                  type: string
                version:
                  type: string
                  nullable: true
                activate:
                  type: boolean
                  default: false
      responses:
        "200":
          description: "注册成功"
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /admin/prompts/{name}/active:
    put:
      operationId: changeActiveVersion
      summary: "切换活跃版本"
      description: "切换提示词的活跃版本"
      tags: [管理后台]
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [version]
              properties:
                version:
                  type: string
      responses:
        "200":
          description: "切换成功"
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /admin/prompts/{name}/ab-test:
    get:
      operationId: getABTestConfig
      summary: "获取 A/B 测试配置"
      description: "获取指定提示词的 A/B 测试配置"
      tags: [管理后台]
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "A/B 测试配置"
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
    post:
      operationId: startABTest
      summary: "启动 A/B 测试"
      description: "为指定提示词启动 A/B 测试"
      tags: [管理后台]
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [canary_version]
              properties:
                canary_version:
                  type: string
                traffic_fraction:
                  type: number
                  default: 0.1
                  minimum: 0.0
                  maximum: 0.5
      responses:
        "200":
          description: "A/B 测试已启动"
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
    delete:
      operationId: stopABTest
      summary: "停止 A/B 测试"
      description: "停止指定提示词的 A/B 测试"
      tags: [管理后台]
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "A/B 测试已停止"
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /admin/prompts/{name}/ab-results:
    get:
      operationId: getABResults
      summary: "获取 A/B 测试结果"
      description: "获取指定提示词的 A/B 测试聚合结果"
      tags: [管理后台]
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "A/B 测试结果"
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
    post:
      operationId: recordABResult
      summary: "记录 A/B 测试结果"
      description: "记录一条 A/B 测试结果用于后续分析"
      tags: [管理后台]
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [version]
              properties:
                version:
                  type: string
                success:
                  type: boolean
                  default: true
                f1:
                  type: number
                  nullable: true
                latency_ms:
                  type: number
                  nullable: true
      responses:
        "200":
          description: "记录成功"
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true

  # ──────────────────────── 数据大屏 ────────────────────────
...
  /dashboard/trends:
    get:
      operationId: getDashboardTrends
      summary: "大屏趋势数据"
      description: "获取大屏时间序列趋势图表数据"
      tags: [数据大屏]
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [7d, 30d, 90d]
            default: 7d
      responses:
        "200":
          description: "趋势数据"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardTrendsResponse"
...
  /dashboard/realtime-poll:
    get:
      operationId: dashboardRealtimePoll
      summary: "大屏实时事件轮询"
      description: "SSE 轮询回退，返回最近事件数组"
      tags: [数据大屏]
      parameters:
        - name: since
          in: query
          schema:
            type: number
          description: "Unix 时间戳，仅返回此时间之后的事件"
      responses:
        "200":
          description: "最近事件"
          content:
            application/json:
              schema:
                type: object
                required: [events, poll_interval_ms]
                properties:
                  events:
                    type: array
                    items:
                      type: object
                  poll_interval_ms:
                    type: integer
                    example: 5000

  # ──────────────────────── 学习中心 ────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT Bearer token obtained from authentication endpoint"
  schemas:
...
    Error:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          description: "错误描述"
          example: "岗位不存在"
        code:
          type: string
          description: "错误码（可选，部分端点返回）"
          example: "POSITION_NOT_FOUND"
        timestamp:
          type: string
          format: date-time
          description: "错误发生时间（可选）"

    # ──────────────────────── 数据大屏 ────────────────────────
...
    DashboardTrendsResponse:
      type: object
      required: [period, data_points, summary]
      properties:
        period:
          type: string
        data_points:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              total_records:
                type: integer
              new_records:
                type: integer
              quality_score:
                type: number
              extractions:
                type: integer
        summary:
          type: object
...